Group sparse Correspondence Analysis

Group sparse Correspondence Analysis

Usage

sparseCA(
  DATA,
  components = 0,
  tol = .Machine$double.eps,
  doublecentering = TRUE,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, components),
  rdsRight = rep(1, components),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)

Arguments

DATA

The contingency table for analysis.

components

The number of dimensions for the analysis; defaults to 0.

tol

Tolerance for the convergence criterion; defaults to .Machine$double.eps.

doublecentering

Logical indicating whether double centering should be applied; defaults to TRUE.

init

Method for initialization; defaults to "svd".

initLeft

Initial values for the left side; defaults to NULL.

initRight

Initial values for the right side; defaults to NULL.

seed

Seed for random number generation; defaults to NULL.

rdsLeft

Radii for the left side; defaults to rep(1, components).

rdsRight

Radii for the right side; defaults to rep(1, components).

grpLeft

Group assignments for the left side; defaults to NULL.

grpRight

Group assignments for the right side; defaults to NULL.

orthogonality

Type of orthogonality constraint; defaults to "loadings".

OrthSpaceLeft

Orthogonal space for the left side; defaults to NULL.

OrthSpaceRight

Orthogonal space for the right side; defaults to NULL.

projPriority

Priority of the projection; defaults to "orth".

projPriorityLeft

Priority of the left projection; defaults to projPriority.

projPriorityRight

Priority of the right projection; defaults to projPriority.

itermaxALS

Maximum number of ALS iterations; defaults to 1000.

itermaxPOCS

Maximum number of POCS iterations; defaults to 1000.

epsALS

Convergence criterion for ALS; defaults to 1e-10.

epsPOCS

Convergence criterion for POCS; defaults to 1e-10.

Value

Returns an object of class c("sSVD", "sGSVD", "list") containing the results of the group sparse correspondence analysis. This object includes various components produced by the sparse generalized singular value decomposition (sparseGSVD) and the subsequent spafac.out processing. Key components include the singular values, left and right singular vectors, and additional metrics relevant to the analysis. The object also contains X.preproc, the preprocessed data matrix used in the analysis.

Examples

data("example1_sCA") # Load the Cause of Deaths Data
sca.res <- sparseCA(as.matrix(example1_sCA), components = 2L, rdsLeft = c(1.3, 1.3), rdsRight = c(2.3, 2.3))
#> [1] 0.4520297 0.1799586