Sparse Multidimensional Scaling

Performs sparse multidimensional scaling (MDS) on a given data matrix.

Usage

sparseMDS(
  DATA,
  masses = NULL,
  components = 0,
  tol = .Machine$double.eps,
  method = "distance",
  init = "svd",
  seed = NULL,
  grp = NULL,
  rds = rep(1, components),
  orthogonality = "loadings",
  OrthSpace = NULL,
  projPriority = "orth",
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)

Arguments

DATA

The data matrix for MDS analysis.

masses

A vector of masses for the rows/columns; defaults to NULL with equal masses of 1/(number of rows).

components

The number of dimensions for the analysis; defaults to 0.

tol

Tolerance for the convergence criterion; defaults to .Machine$double.eps.

method

Method for analysis, either "distance" or "covariance".

init

Initialization method; defaults to "svd".

seed

Seed for random number generation; defaults to NULL.

rds

Radii for the analysis; defaults to rep(1, components).

orthogonality

Type of orthogonality constraint; defaults to "loadings".

OrthSpace

Orthogonal space for the analysis; defaults to NULL.

projPriority

Priority of the projection; defaults to "orth".

itermaxALS

Maximum number of ALS iterations; defaults to 1000.

itermaxPOCS

Maximum number of POCS iterations; defaults to 1000.

epsALS

Convergence criterion for ALS; defaults to 1e-10.

epsPOCS

Convergence criterion for POCS; defaults to 1e-10.

Value

Returns an object containing the results of the sparse MDS analysis, including eigenvalues, eigenvectors, and various metrics relevant to the analysis.

Examples

# Example usage of sparseMDS function
# Assuming `data` is a data matrix
#' \dontrun{result <- sparseMDS(DATA = data)}