sparse Multiple Factor Analysis

sparse Multiple Factor Analysis

Usage

sparseMFA(
  X,
  column.design,
  components = 0,
  sparseOption = "variable",
  center = TRUE,
  scale = TRUE,
  mfa.scale = TRUE,
  tol = .Machine$double.eps,
  init = "svd",
  initLeft = NULL,
  initRight = NULL,
  seed = NULL,
  rdsLeft = rep(1, components),
  rdsRight = rep(1, components),
  grpLeft = NULL,
  grpRight = NULL,
  orthogonality = "loadings",
  OrthSpaceLeft = NULL,
  OrthSpaceRight = NULL,
  projPriority = "orth",
  projPriorityLeft = projPriority,
  projPriorityRight = projPriority,
  itermaxALS = 1000,
  itermaxPOCS = 1000,
  epsALS = 1e-10,
  epsPOCS = 1e-10
)

Source

Some lines of the function are inspired by the MExPosition package by Derek Beaton and Cherise Chin Fatt.

Arguments

X

Data matrix with I rows and J columns

components

the number of dimensions to extract

sparseOption

"variable" or "subtable"

center

For the X matrix: A parameter to pass through to center in scale function; either a logical value or numeric-alike vector of length equal to the number of columns of X.

scale

For the X matrix: A parameter to pass through to scale in scale function; either a logical value or numeric-alike vector of length equal to the number of columns of X.

tol

default is .Machine$double.eps. A parameter to pass through to gplssvd; eliminates singular values that are effectively zero (and thus drops null components).

epsPOCS